Refactoring with namespaces
Details: as good suggestions came in the PR reviews, now all helpers will live under
`factories/(metric_namespace).py`
